+++ src/engine/core/examples/make_config.rs
@@ -1,5 +1,5 @@
 fn main() {
-    println!(
+    print!(
         "{}",
         serde_yaml::to_string(&kime_engine_core::RawConfig::default()).unwrap()
     );