+++ src/engine/core/src/lib.rs
@@ -1,83 +1,22 @@
-mod characters;
 mod config;
-mod input_result;
-mod keycode;
-mod state;
-
 mod os;
 
-use ahash::AHashMap;
+use config::{HotkeyBehavior, HotkeyResult, IconColor};
 
-use crate::characters::KeyValue;
-use crate::os::{DefaultOsContext, OsContext};
-use enumset::EnumSet;
+use os::{DefaultOsContext, OsContext};
 
-pub use crate::config::{
-    Addon, Config, Hotkey, HotkeyBehavior, HotkeyResult, IconColor, InputCategory, RawConfig,
-    BUILTIN_LAYOUTS,
-};
-pub use crate::input_result::InputResult;
-pub use crate::keycode::{Key, KeyCode, ModifierState};
-pub use crate::state::HangulState;
+use kime_engine_backend::InputEngineBackend;
+use kime_engine_backend_hangul::{HangulConfig, HangulEngine};
+use kime_engine_backend_latin::{LatinConfig, LatinEngine};
 
-pub struct LayoutContext {
-    pub input_category: InputCategory,
-    pub layout_addons: EnumSet<Addon>,
-}
+pub use config::{Config, Hotkey, InputCategory, RawConfig};
 
-impl Default for LayoutContext {
-    fn default() -> Self {
-        Self::new(&Config::default())
-    }
-}
-
-impl LayoutContext {
-    pub fn new(config: &Config) -> Self {
-        Self {
-            input_category: config.default_category,
-            layout_addons: config.layout_addons[config.default_category],
-        }
-    }
-
-    pub fn update_layout(&mut self, category: InputCategory, config: &Config) {
-        self.input_category = category;
-        self.layout_addons = config.layout_addons[category];
-    }
-
-    pub fn check_addon(&self, addon: Addon) -> bool {
-        self.layout_addons.contains(addon)
-    }
-}
-
-#[derive(Clone, Default)]
-pub struct Layout {
-    keymap: AHashMap<Key, KeyValue>,
-}
-
-impl Layout {
-    fn from_items(items: AHashMap<Key, String>) -> Self {
-        let mut keymap = AHashMap::new();
-
-        for (key, value) in items {
-            let value = match value.parse::<KeyValue>() {
-                Ok(value) => value,
-                Err(_) => continue,
-            };
-
-            keymap.insert(key, value);
-        }
-
-        Self { keymap }
-    }
-
-    pub fn load_from(content: &str) -> Result<Self, serde_yaml::Error> {
-        Ok(Self::from_items(serde_yaml::from_str(content)?))
-    }
-}
+pub use kime_engine_backend::{InputResult, Key, KeyCode, ModifierState};
 
 pub struct InputEngine {
-    state: HangulState,
-    layout_ctx: LayoutContext,
+    engine_impl: EngineImpl,
+    commit_buf: String,
+    preedit_buf: String,
     os_ctx: DefaultOsContext,
     icon_color: IconColor,
 }
@@ -91,29 +30,27 @@ impl Default for InputEngine {
 impl InputEngine {
     pub fn new(config: &Config) -> Self {
         Self {
-            state: HangulState::new(config.word_commit),
+            engine_impl: EngineImpl::new(config),
+            commit_buf: String::with_capacity(16),
+            preedit_buf: String::with_capacity(16),
             os_ctx: DefaultOsContext::default(),
-            layout_ctx: LayoutContext::new(config),
             icon_color: config.icon_color,
         }
     }
 
-    pub fn set_input_category(&mut self, config: &Config, category: InputCategory) {
-        self.layout_ctx.update_layout(category, config);
+    pub fn set_input_category(&mut self, category: InputCategory) {
+        // Reset previous engine
+        self.engine_impl.clear_preedit(&mut self.commit_buf);
+        self.engine_impl.category = category;
     }
 
     pub fn category(&self) -> InputCategory {
-        self.layout_ctx.input_category
+        self.engine_impl.category
     }
 
     pub fn update_layout_state(&mut self) -> std::io::Result<()> {
         self.os_ctx
-            .update_layout_state(self.layout_ctx.input_category, self.icon_color)
-    }
-
-    fn bypass(&mut self) -> InputResult {
-        self.clear_preedit();
-        InputResult::NEED_RESET
+            .update_layout_state(self.category(), self.icon_color)
     }
 
     fn try_get_global_input_category_state(&mut self, config: &Config) {
@@ -123,8 +60,8 @@ impl InputEngine {
                 .read_global_hangul_state()
                 .unwrap_or(self.category());
 
-            if self.layout_ctx.input_category != global {
-                self.layout_ctx.update_layout(global, config);
+            if self.category() != global {
+                self.set_input_category(global);
             }
         }
     }
@@ -136,76 +73,73 @@ impl InputEngine {
 
             match hotkey.behavior() {
                 HotkeyBehavior::Switch(category) => {
-                    if self.layout_ctx.input_category != category {
-                        self.layout_ctx.update_layout(category, config);
+                    if self.category() != category {
+                        self.set_input_category(category);
                         ret |= InputResult::LANGUAGE_CHANGED;
                         processed = true;
                     }
                 }
                 HotkeyBehavior::Toggle(left, right) => {
-                    let change = if self.layout_ctx.input_category == left {
+                    let change = if self.category() == left {
                         right
-                    } else if self.layout_ctx.input_category == right {
+                    } else if self.category() == right {
                         left
                     } else {
                         right
                     };
 
-                    self.layout_ctx.update_layout(change, config);
+                    self.set_input_category(change);
                     ret |= InputResult::LANGUAGE_CHANGED;
                     processed = true;
                 }
                 HotkeyBehavior::Emoji => {
-                    if self.os_ctx.emoji(&mut self.state).unwrap_or(false) {
-                        ret |= InputResult::NEED_RESET;
+                    if self
+                        .os_ctx
+                        .emoji(&mut self.engine_impl, &mut self.commit_buf)
+                        .is_ok()
+                    {
                         processed = true;
                     }
                 }
                 HotkeyBehavior::Hanja => {
-                    if self.os_ctx.hanja(&mut self.state).unwrap_or(false) {
-                        ret |= InputResult::NEED_RESET;
+                    if self
+                        .os_ctx
+                        .hanja(&mut self.engine_impl, &mut self.commit_buf)
+                        .is_ok()
+                    {
                         processed = true;
                     }
                 }
                 HotkeyBehavior::Commit => {
-                    if self
-                        .state
-                        .preedit_result()
-                        .contains(InputResult::HAS_PREEDIT)
-                    {
-                        self.state.clear_preedit();
-                        ret |= InputResult::NEED_RESET;
+                    if self.engine_impl.has_preedit() {
+                        self.engine_impl.clear_preedit(&mut self.commit_buf);
                         processed = true;
                     }
                 }
             }
 
             match (hotkey.result(), processed) {
-                (HotkeyResult::Bypass, _) | (HotkeyResult::ConsumeIfProcessed, false) => {
-                    ret |= self.bypass();
-                }
+                (HotkeyResult::Bypass, _) | (HotkeyResult::ConsumeIfProcessed, false) => {}
                 (HotkeyResult::Consume, _) | (HotkeyResult::ConsumeIfProcessed, true) => {
-                    ret |= InputResult::CONSUMED | self.state.preedit_result();
+                    ret |= InputResult::CONSUMED;
                 }
             }
 
+            ret |= self.current_result();
+
             ret
-        } else if key.code == KeyCode::Shift {
-            // Don't reset state
-            self.state.preedit_result()
         } else {
             self.try_get_global_input_category_state(config);
 
-            if key.code == KeyCode::Backspace {
-                self.state.backspace(&self.layout_ctx)
-            } else if let Some(v) = config.layouts[self.layout_ctx.input_category]
-                .keymap
-                .get(&key)
-            {
-                self.state.key(v, &self.layout_ctx)
-            } else {
-                self.bypass()
+            let mut ret = InputResult::empty();
+
+            if self.engine_impl.press_key(key, &mut self.commit_buf) {
+                ret |= InputResult::CONSUMED;
             }
+
+            ret |= self.current_result();
+
+            ret
         }
     }
 
@@ -217,32 +151,101 @@ impl InputEngine {
     ) -> InputResult {
         match KeyCode::from_hardward_code(hardware_code) {
             Some(code) => self.press_key(Key::new(code, state), config),
-            None => self.bypass(),
+            None => self.current_result(),
         }
     }
 
     #[inline]
+    pub fn clear_commit(&mut self) {
+        self.commit_buf.clear();
+    }
+
+    #[inline]
     pub fn clear_preedit(&mut self) {
-        self.state.clear_preedit();
+        self.engine_impl.clear_preedit(&mut self.commit_buf);
+    }
+
+    #[inline]
+    pub fn remove_preedit(&mut self) {
+        self.engine_impl.reset();
     }
 
     #[inline]
     pub fn preedit_str(&mut self) -> &str {
-        self.state.preedit_str()
+        self.preedit_buf.clear();
+        self.engine_impl.preedit_str(&mut self.preedit_buf);
+        &self.preedit_buf
     }
 
     #[inline]
-    pub fn commit_str(&mut self) -> &str {
-        self.state.commit_str()
-    }
-
-    #[inline]
-    pub fn flush(&mut self) {
-        self.state.flush();
+    pub fn commit_str(&self) -> &str {
+        &self.commit_buf
     }
 
     #[inline]
     pub fn reset(&mut self) {
-        self.state.reset();
+        self.clear_commit();
+        self.remove_preedit();
+    }
+
+    fn current_result(&self) -> InputResult {
+        let mut ret = InputResult::empty();
+        if self.engine_impl.has_preedit() {
+            ret |= InputResult::HAS_PREEDIT;
+        }
+        if !self.commit_buf.is_empty() {
+            ret |= InputResult::HAS_COMMIT;
+        }
+        ret
+    }
+}
+
+struct EngineImpl {
+    category: InputCategory,
+    latin_engine: LatinEngine,
+    hangul_engine: HangulEngine,
+    math_engine: LatinEngine,
+}
+
+impl EngineImpl {
+    pub fn new(config: &Config) -> Self {
+        Self {
+            category: config.default_category,
+            latin_engine: config.latin_engine.clone(),
+            hangul_engine: config.hangul_engine.clone(),
+            math_engine: config.latin_engine.clone(),
+        }
+    }
+}
+
+macro_rules! do_engine {
+    ($self:expr, $func:ident($($arg:expr),*)) => {
+        match $self.category {
+            InputCategory::Hangul => $self.hangul_engine.$func($($arg,)*),
+            InputCategory::Latin => $self.latin_engine.$func($($arg,)*),
+            InputCategory::Math => $self.math_engine.$func($($arg,)*),
+        }
+    };
+}
+
+impl InputEngineBackend for EngineImpl {
+    fn press_key(&mut self, key: Key, commit_buf: &mut String) -> bool {
+        do_engine!(self, press_key(key, commit_buf))
+    }
+
+    fn clear_preedit(&mut self, commit_buf: &mut String) {
+        do_engine!(self, clear_preedit(commit_buf));
+    }
+
+    fn reset(&mut self) {
+        do_engine!(self, reset());
+    }
+
+    fn has_preedit(&self) -> bool {
+        do_engine!(self, has_preedit())
+    }
+
+    fn preedit_str(&self, buf: &mut String) {
+        do_engine!(self, preedit_str(buf));
     }
 }